Change FrameData's frame_index to uint32_t
This is a step toward unifying various frame counters.

For 120 fps rendering, uint32_t is roughly
  2^32 / (120 * 60 * 60 * 24) ~= 414 days.

More than long enough.

TEST=bots build

Bug: 40578019
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6892151
Auto-Submit: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507263}

+++ components/viz/service/display/frame_interval_decider_unittest.cc
@@ -122,7 +122,7 @@ class FrameIntervalDeciderTest : public testing::Test,
 
   void UpdateFrame(Surface* surface,
                    FrameIntervalInputs frame_interval_inputs) {
-    uint64_t frame_index = surface->GetActiveFrameIndex() + 1u;
+    uint32_t frame_index = surface->GetActiveFrameIndex() + 1u;
     auto frame = MakeDefaultCompositorFrame();
     frame.metadata.frame_interval_inputs = std::move(frame_interval_inputs);
     ASSERT_TRUE(surface->QueueFrame(std::move(frame), frame_index,
